# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Berita Documentation",
        "description": "Dokumentasi API untuk sistem manajemen berita",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/likes": {
            "get": {
                "tags": [
                    "Disukai"
                ],
                "summary": "Get all likes",
                "description": "Mengambil semua like/disukai (umum, tidak memerlukan login)",
                "operationId": "65e88aa962ee2b93499d2cc54bb5a8ff",
                "responses": {
                    "200": {
                        "description": "List semua like",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Disukai"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Disukai"
                ],
                "summary": "Create or toggle like",
                "description": "Menambahkan like baru atau update jika sudah ada. Hanya dapat dilakukan oleh user yang sedang login.",
                "operationId": "6b998b51fbe77ebda9a4193f66f61630",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_berita"
                                ],
                                "properties": {
                                    "id_berita": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "suka": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Like berhasil dibuat atau diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Disukai"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/likes/{id_disukai}": {
            "put": {
                "tags": [
                    "Disukai"
                ],
                "summary": "Update like (only owner)",
                "description": "Mengubah status like/disukai. Hanya user pemilik like yang dapat mengubah.",
                "operationId": "6326b661e39527f636abf765ff9ec306",
                "parameters": [
                    {
                        "name": "id_disukai",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "suka"
                                ],
                                "properties": {
                                    "suka": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Like berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Disukai"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Akses ditolak (bukan pemilik like)"
                    },
                    "404": {
                        "description": "Like tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Disukai"
                ],
                "summary": "Delete like (only owner)",
                "description": "Menghapus like/disukai. Hanya user pemilik like yang dapat menghapus.",
                "operationId": "db764387b1180ecd433ed3da28e6a41c",
                "parameters": [
                    {
                        "name": "id_disukai",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Like berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Akses ditolak (bukan pemilik like)"
                    },
                    "404": {
                        "description": "Like tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Disukai": {
                "title": "Disukai",
                "description": "Data Like/Disukai",
                "properties": {
                    "id_disukai": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_user": {
                        "type": "integer",
                        "example": 3
                    },
                    "id_berita": {
                        "type": "integer",
                        "example": 7
                    },
                    "suka": {
                        "description": "True jika like, false jika dislike",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-06T08:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-06T08:10:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter your Bearer token",
                "bearerFormat": "JWT"
            }
        }
    },
...
}